feat: initial analysis-manager-kit package implementation
