+++ packages/server-analysis-manager-kit/src/components/builder/types.ts
