+++ packages/server-analysis-manager-kit/src/components/core/constants.ts
