+++ packages/server-analysis-manager-kit/src/components/core/types.ts
