+++ packages/server-analysis-manager-kit/src/index.ts
@@ -1,12 +1,10 @@
 /*
- * Copyright (c) 2023-2024.
+ * Copyright (c) 2024.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './commands';
-export * from './helpers';
-export * from './error';
-export * from './module';
-export * from './utils';
+export * from './components';
+export * from './constants';
+export * from './types';
