+++ packages/server-analysis-manager-kit/src/types.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2024.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+import type { ObjectLiteral } from '@privateaim/server-kit';
+
+export type QueueRouterPayload<T extends ObjectLiteral = ObjectLiteral> = {
+    data: T,
+    metadata: {
+        command: string,
+        event?: string,
+        component: string
+    }
+};
