+++ packages/server-analysis-manager-kit/tsconfig.build.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig.build.json",
+
+    "compilerOptions": {
+        "outDir": "./dist"
+    },
+
+    "include": [
+        "src/**/*.ts"
+    ]
+}
