+++ packages/server-analysis-manager/src/components/core/index.ts
@@ -6,8 +6,6 @@
  */
 
 export * from './commands';
-export * from './constants';
 export * from './error';
 export * from './module';
-export * from './type';
 export * from './utils';
