+++ packages/server-analysis-manager/src/components/index.ts
@@ -7,6 +7,5 @@
 
 export * from './builder';
 export * from './core';
-export * from './constants';
 export * from './type';
 export * from './utils';
