+++ packages/server-analysis-manager/tsconfig.json
@@ -10,6 +10,9 @@
             ],
             "@privateaim/storage-kit": [
                 "./packages/storage-kit/src"
+            ],
+            "@privateaim/server-analysis-manager-kit": [
+                "./packages/server-analysis-manager-kit/src"
             ]
         }
     }
