+++ packages/server-core/package.json
@@ -21,7 +21,7 @@
         "@privateaim/core": "^0.4.0",
         "@privateaim/server-kit": "^0.4.0",
         "@privateaim/server-http-kit": "^0.4.0",
-        "@privateaim/server-analysis-manager": "^0.4.0",
+        "@privateaim/server-analysis-manager-kit": "^0.4.0",
         "@routup/basic": "^1.3.1",
         "@routup/decorators": "^3.3.1",
         "@routup/prometheus": "^2.3.0",
