+++ packages/server-core/src/domains/analysis/commands/build-status.ts
@@ -5,7 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { BuilderCommand, buildBuilderQueuePayload } from '@privateaim/server-analysis-manager';
+import { BuilderCommand, buildBuilderQueuePayload } from '@privateaim/server-analysis-manager-kit';
 import { useDataSource } from 'typeorm-extension';
 import { useAmqpClient } from '@privateaim/server-kit';
 import { resolveAnalysis } from './utils';
