+++ packages/server-core/src/http/controllers/core/analysis/handlers/delete.ts
@@ -8,7 +8,7 @@
 import { ForbiddenError, NotFoundError } from '@ebec/http';
 import { PermissionID } from '@privateaim/core';
 import { isRealmResourceWritable } from '@authup/core-kit';
-import { CoreCommand, buildCoreQueuePayload } from '@privateaim/server-analysis-manager';
+import { CoreCommand, buildCoreQueuePayload } from '@privateaim/server-analysis-manager-kit';
 import type { Request, Response } from 'routup';
 import { sendAccepted, useRequestParam } from 'routup';
 import { useDataSource } from 'typeorm-extension';
